Add event handler for numbers and operations.
- Changed overall manner of calculator layout generation. Previous
model used DOM manipulation method to create grid for buttons. Now
the layout in it's entirety is created through HTML elements and CSS
flexbox. The former made harder the job of handling buttons, because
each button has an id-level of functionality. To achieve it, I'd need
to create several conditionals to change each button id and classes,
making the code a mess. This way, targeting nodes is fast and
straightfoward;

- Calculator has now primitive event handlers for numbers, equal sign
and operators. It's now somewhat functional, but has several bugs to
be handled, such as:
    .Pressing operators before entering numbers make inputArr to store
    undefined values;
    .Pressing equal sign out of a calculator normal input flow causes
    undesirable result;
    .Particular cases are not yet handled, such as division by zero;
    .Del and Clear are not yet coded;

For further updates, I suggest adding a log of input for the UI.